+++ tree_shaking/config.py
@@ -1,13 +1,14 @@
 import atexit
 import hashlib
+import os
+import sys
 import typing as t
+from functools import cache
 from functools import partial
-from os.path import isabs
-
 from lk_utils import fs
-
+from lk_utils import run_cmd_args
+from os.path import isabs
 from .path_scope import path_scope
-
 
 class T:
     AnyDirPath = str
@@ -25,15 +26,17 @@ class T:
     #       pillow      pil
     NormPath = str  # absolute path.
     RelPath = str  # relative path, starts from `root`.
+    SpecialPath = str  # '$venv' or `$venv/...`
     
     # noinspection PyTypedDict
     Config0 = t.TypedDict('Config0', {
         'root'        : AnyDirPath,
-        'search_paths': t.List[RelPath],
+        'search_paths': t.List[t.Union[RelPath, SpecialPath]],
         'entries'     : t.List[RelPath],  # must ends with ".py"
         'ignores'     : t.List[IgnoredName],
-        'export'      : t.Optional[t.TypedDict('SoleExportOption', {
-            'source': t.Optional[AnyDirPath], 'target': AnyDirPath,
+        'export'      : t.Optional[t.TypedDict('ExportOption0', {
+            'source': t.Union[SpecialPath, AnyDirPath],
+            'target': AnyDirPath,
         })],
     }, total=False)
     """
@@ -52,16 +55,14 @@ class T:
         'search_paths': t.List[NormPath],
         'entries'     : t.Dict[NormPath, GraphId],
         'ignores'     : t.Union[t.FrozenSet[str], t.Tuple[str, ...]],
-        'export'      : t.TypedDict('SoleExportOption', {
+        'export'      : t.TypedDict('ExportOption1', {
             'source': NormPath, 'target': NormPath,
         }),
     })
     
     Config = Config1
 
-
 graphs_root = fs.xpath('_cache/module_graphs')
-
 
 def parse_config(file: str, _save: bool = False, **kwargs) -> T.Config:
     """
@@ -84,7 +85,7 @@ def parse_config(file: str, _save: bool = False, **kwargs) -> T.Config:
     }
     
     # 1
-    if isabs(cfg0['root']):
+    if isabs(cfg0['root']):  # not suggested
         cfg1['root'] = fs.normpath(cfg0['root'])
     else:
         cfg1['root'] = fs.normpath('{}/{}'.format(cfg_dir, cfg0['root']))
@@ -92,8 +93,13 @@ def parse_config(file: str, _save: bool = False, **kwargs) -> T.Config:
     # 2
     _root = cfg1['root']
     
-    def fmtpath(p: T.RelPath) -> T.NormPath:
-        if p == '.': return _root
+    def fmtpath(p: t.Union[T.RelPath, T.SpecialPath]) -> T.NormPath:
+        if p == '': 
+            raise ValueError('path cannot be empty')
+        if p == '.': 
+            return _root
+        if p.startswith('$venv'): 
+            return p.replace('$venv', _get_venv_root(_root), 1)
         assert not p.startswith(('./', '../', '<')), p
         out = '{}/{}'.format(_root, p)
         assert fs.exist(out), out
@@ -111,19 +117,17 @@ def parse_config(file: str, _save: bool = False, **kwargs) -> T.Config:
         temp[p] = hash_path_to_uid(p)
     
     # 4
-    cfg1['ignores'] = frozenset(cfg0.get('ignores', ()))
+    cfg1['ignores'] = frozenset(cfg0.get('ignores', ()))  # type: ignore
     
     # 5
     dict0 = kwargs.get('export', {'source': '', 'target': ''})
     dict1 = cfg0.get('export', {'source': '', 'target': ''})
     if src := (dict0['source'] or dict1['source']):
-        assert src in cfg0['search_paths']  # noqa
+        assert src in cfg0['search_paths']
         cfg1['export']['source'] = fmtpath(src)
-        # cfg1['export']['source'] = src
     if dict0['target']:
         cfg1['export']['target'] = fs.abspath(dict0['target'])
     elif dict1['target']:
-        # cfg1['export']['target'] = fmtpath(dict1['target'])
         cfg1['export']['target'] = fs.normpath('{}/{}'.format(
             cfg1['root'], dict1['target']
         ))
@@ -134,10 +138,37 @@ def parse_config(file: str, _save: bool = False, **kwargs) -> T.Config:
     # print(cfg1, ':l')
     return cfg1
 
-
 def hash_path_to_uid(abspath: str) -> str:
     return hashlib.md5(abspath.encode()).hexdigest()
 
+@cache
+def _get_venv_root(working_root: str) -> T.NormPath:
+    """
+    find venv root (the "site-packages" folder) by `poetry env` command.
+    """
+    # https://stackoverflow.com/questions/75232761/
+    if 'VIRTUAL_ENV' in os.environ:
+        del os.environ['VIRTUAL_ENV']
+    venv_root = fs.normpath(
+        run_cmd_args(
+            (
+                sys.executable, '-m', 'poetry', 'env', 'info', '--path', 
+                '--no-ansi', '--directory', working_root
+            ), 
+            cwd=working_root,
+        )
+    )
+    print(venv_root)
+    assert venv_root.endswith('py3.12')
+    
+    if os.name == 'nt':
+        out = '{}/Lib/site-packages'.format(venv_root)
+    else:
+        out = '{}/lib/python{}.{}/site-packages'.format(
+            venv_root, sys.version_info.major, sys.version_info.minor
+        )
+    assert fs.exist(out), (working_root, venv_root, out)
+    return out
 
 def _save_graph_alias(config: T.Config1) -> None:
     map_ = fs.load(fs.xpath('_cache/module_graphs_alias.yaml'), default={})